backend/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
sqlalchemy[asyncio]==2.0.36
asyncpg==0.30.0
psycopg2-binary==2.9.10
alembic==1.14.1
httpx==0.28.1
feedparser==6.0.11
anthropic==0.43.0
openai==1.59.3
google-genai==1.5.0
apscheduler~=3.10
pydantic-settings==2.7.1
python-jose[cryptography]==3.3.0
bcrypt==4.2.1
python-dateutil==2.9.0
